Experimenting with passing not primitive parameters to jetpack navigation. We have LocaleDate which isn't a primitive type so we need to Serialize it and also have the Diary as parcelable to use the NavCustomArgHelper.
We also lose the type safe if we want to get the nav argument from the savedinstancestate of the viewmodel for that we need so save the argument value name and use it to get it from saved state.
toRoute() doesn't work in this cases from the savedinstancestate of the viewmodel.